@@ -11,9 +11,9 @@
     "chart.js": "^3.6.2",
     "leaflet": "^1.9.3",
     "lodash": "^4.17.21",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "react-chartjs-2": "^4.0.0",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.2.0",
     "react-leaflet": "^4.2.0",
     "react-scripts": "5.0.0",
     "web-vitals": "^2.1.2"
